operator: extract CR-agnostic Broker machinery into internal/brokerset
Move the Broker CR management core out of the V1-coupled
pkg/resources/brokerset.go into a new operator/internal/brokerset package:
rendering Broker CRs from a desired StatefulSet, steady-state
reconciliation (create/update, scale-down via decommission intent),
cluster-wide roll-grant serialization, the STS->Broker migration state
machine, and rollback.

Everything owning-CR-specific is injected via the BrokerSet config: owner
object, ClusterRef, label sets/selectors, the config-checksum annotation
key (V1 configmap-hash vs V2 config.redpanda.com/checksum), and callbacks
for admin-health gating, quiescence (V1: clear Restarting), migration
blockers (V1: restarting/decommissioning status), and migration progress
reporting (V1: BrokerMigration condition). This prepares the V2 Redpanda
reconciler to reuse the exact same machinery and invariants.

Behavior-preserving changes only:
- pod Subdomain now comes from sts.Spec.ServiceName instead of
  cluster.Name (identical for V1, which sets ServiceName: cluster.Name)
- a live STS with a nil desired render no longer enters the migration
  path (previously an unreachable nil-cast panic)

RequeueAfterError moves to brokerset with a type alias left in
pkg/resources so existing errors.As call sites keep matching.
BrokerSetResource stays as a thin V1 adapter and the unit-test surface;
pkg/resources tests pass unchanged.
