+++ operator/pkg/resources/statefulset_update.go
@@ -36,6 +36,7 @@ import (
 	k8sclient "sigs.k8s.io/controller-runtime/pkg/client"
 
 	vectorizedv1alpha1 "github.com/redpanda-data/redpanda-operator/operator/api/vectorized/v1alpha1"
+	"github.com/redpanda-data/redpanda-operator/operator/internal/brokerset"
 	"github.com/redpanda-data/redpanda-operator/operator/pkg/labels"
 	"github.com/redpanda-data/redpanda-operator/operator/pkg/utils"
 )
@@ -923,19 +924,11 @@ func (r *StatefulSetResource) evaluateUnderReplicatedPartitions(
 	return nil
 }
 
-// RequeueAfterError error carrying the time after which to requeue.
-type RequeueAfterError struct {
-	RequeueAfter time.Duration
-	Msg          string
-}
-
-func (e *RequeueAfterError) Error() string {
-	return fmt.Sprintf("RequeueAfterError %s", e.Msg)
-}
-
-func (e *RequeueAfterError) Is(target error) bool {
-	return e.Error() == target.Error()
-}
+// RequeueAfterError error carrying the time after which to requeue. The
+// canonical definition lives in operator/internal/brokerset (the CR-agnostic
+// Broker machinery produces these too); the alias keeps type identity for
+// every errors.As check against *resources.RequeueAfterError.
+type RequeueAfterError = brokerset.RequeueAfterError
 
 // RequeueError error to requeue using default retry backoff.
 type RequeueError struct {
